move to helpers and use clierror
